split table into component
